services/aggregation/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
gunicorn>=20.0,<21.0
psycopg2-binary>=2.8,<3.0
django-cors-headers>=3.7,<4.0
django-environ>=0.4.5,<0.5
pyspark